added components to the dashboard by clicking buttons, moved products contex to Context folder
