+++ client/src/Contexts/ProductsContext.js
@@ -1,5 +1,3 @@
-
-
 import React, { createContext, useState } from "react";
 
 export const ProductsContext = createContext();
@@ -22,8 +20,8 @@ const ProductsProvider = ({ children }) => {
      setError(true);
      throw new Error("Network response was not ok.");
    }
- };
- 
+  };
+  
   return (
     <ProductsContext.Provider value={{ products, loading, error, fetchData}}>
       {" "}
