+++ client/src/services/orderService.js
@@ -5,6 +5,7 @@ export function insertOrder(orderObj) {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
+     " Access-Control-Allow-Origin": "*",
     },
     body: JSON.stringify(orderObj),
   })
