+++ client/src/styles/navbar.scss
@@ -1,3 +1,4 @@
+@import './variables/variables'; 
 
   .navbar{
     height: 90px;
@@ -7,7 +8,7 @@
     align-items: center;
     position: relative;
     /* padding: 0.5rem 0rem; */
-    background-color: hsl(0, 0%, 4%);
+    background-color: $bg4;
     color: black;
     z-index: 1;
     cursor: pointer;
