+++ client/src/styles/variables/_variables.scss
@@ -2,7 +2,7 @@
    $bg: #F5F5F5;
    $bg2: #D9D9D9;
    $bg3: #AAAAAA;
-
+   $bg4: rgb(24, 23, 23);
    $blue: #2c97ea;
    $green: #6AB999;
    $green2: #70D981;
